Add upload_to_url task to Rakefile and execute this task in after_script
This task POST's the latest build of Ember.js to a Heroku app, which in turn replaces the latest build in the ember.js downloads section on GitHub.

This task is added as after_script in the travis configuration. If the tests succeed, the after_script is invoked and the latest build of Ember.js is added to the downloads section.